Refactor elixir checker to accept phoenix projects
The elixir checker uses the `mix compile` command. This command
has to be executed from the root of the project containing
the `mix.exs` config file.

The commit allow the checker to find this root directory and launch
the command from here, giving access to all libs dependencies.

This is kind of a hack. An issue exist in the flycheck project:
https://github.com/flycheck/flycheck/issues/312